+++ examples/debugging/src/lib.rs
@@ -4,7 +4,7 @@ mod tests {
 
     use fuel_abi_types::program_abi::ProgramABI;
     use fuels::{
-        core::{function_selector::resolve_fn_selector, traits::Parameterize},
+        core::{codec::resolve_fn_selector, traits::Parameterize},
         types::{errors::Result, param_types::ParamType, SizedAsciiString},
     };
