+++ packages/fuels-core/src/codec/abi_decoder.rs
@@ -32,7 +32,7 @@ impl ABIDecoder {
     /// # Examples
     ///
     /// ```
-    /// use fuels_core::abi_decoder::ABIDecoder;
+    /// use fuels_core::codec::ABIDecoder;
     /// use fuels_types::{Token, param_types::ParamType};
     ///
     /// let tokens = ABIDecoder::decode(&[ParamType::U8, ParamType::U8], &[0,0,0,0,0,0,0,1,0,0,0,0,0,0,0,2]).unwrap();
