+++ packages/fuels-programs/src/receipt_parser.rs
@@ -1,6 +1,6 @@
 use fuel_tx::{ContractId, Receipt};
 use fuels_core::{
-    abi_decoder::ABIDecoder,
+    codec::ABIDecoder,
     types::{
         bech32::Bech32ContractId,
         errors::{error, Error, Result},
