+++ packages/fuels-test-helpers/Cargo.toml
@@ -32,5 +32,5 @@ which = { workspace = true, default-features = false }
 
 [features]
 default = ["fuels-accounts", "std"]
-std = ["fuels-accounts?/std"]
+std = ["fuels-accounts?/std", "fuels-core/std"]
 fuel-core-lib = ["fuel-core"]
