+++ packages/wasm-tests/src/lib.rs
@@ -3,7 +3,7 @@ extern crate alloc;
 #[cfg(test)]
 mod tests {
     use fuels::{
-        core::{abi_encoder::ABIEncoder, traits::Tokenizable},
+        core::{codec::ABIEncoder, traits::Tokenizable},
         macros::wasm_abigen,
     };
     use wasm_bindgen_test::wasm_bindgen_test;
